+++ src/NonAligningTransformer.hpp
@@ -11,6 +11,7 @@ public:
     NonAlignedDynamicTransformationElement(const std::string& sourceFrame, const std::string& targetFrame);
     virtual ~NonAlignedDynamicTransformationElement() {};
     virtual bool getTransformation(const base::Time& atTime, bool doInterpolation, TransformationType& result);
+    virtual bool isStatic() { return false; }
 
     void setTransformation(const base::Time& atTime, const TransformationType& tr);
     virtual void setTransformationChangedCallback(boost::function<void (const base::Time &ts)> callback)
